# backend/src/astro_canvas/server/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "EdgeDoc": {
        "description": "``{\"from\": [node, port], \"to\": [node, port]}``.",
        "properties": {
          "from": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "From",
            "type": "array"
          },
          "to": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "string"
              },
              {
                "type": "string"
              }
            ],
            "title": "To",
            "type": "array"
          }
        },
        "required": [
          "from",
          "to"
        ],
        "title": "EdgeDoc",
        "type": "object"
      },
...
      "GroupDoc": {
        "additionalProperties": true,
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "nodes": {
            "items": {
              "type": "string"
            },
            "title": "Nodes",
            "type": "array"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "title": "GroupDoc",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LayoutIssue": {
        "description": "One problem with a layout section, addressed by layout name and item position.",
        "properties": {
          "code": {
            "enum": [
              "bad_layout",
              "bad_ref",
              "unknown_promoted",
              "unknown_view",
              "unknown_node",
              "duplicate_view"
            ],
            "title": "Code",
            "type": "string"
          },
          "index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index"
          },
          "layout": {
            "title": "Layout",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          }
        },
        "required": [
          "layout",
          "code",
          "message"
        ],
        "title": "LayoutIssue",
        "type": "object"
      },
...
      "NodeDoc": {
        "additionalProperties": true,
        "description": "One node instance on the canvas.",
        "properties": {
          "cost": {
            "anyOf": [
              {
                "enum": [
                  "cheap",
                  "expensive",
                  "auto"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "disabled": {
            "default": false,
            "title": "Disabled",
            "type": "boolean"
          },
          "linked": {
            "items": {
              "type": "string"
            },
            "title": "Linked",
            "type": "array"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          },
          "pos": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pos"
          },
          "size": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "ui": {
            "additionalProperties": true,
            "title": "Ui",
            "type": "object"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "type"
        ],
        "title": "NodeDoc",
        "type": "object"
      },
      "NodeIssue": {
        "description": "One compile-time problem on a node (returned per node like ComfyUI's ``node_errors``).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "param": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param"
          },
          "port": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "NodeIssue",
        "type": "object"
      },
...
      "PromotedDoc": {
        "additionalProperties": true,
        "description": "A param lifted out of a node so a layout (or a subgraph instance) can set it.\n\n``ref`` (``\"<node>.<param>\"``) is how layouts and subgraph instances address it.",
        "properties": {
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "order": {
            "default": 0,
            "title": "Order",
            "type": "integer"
          },
          "param": {
            "title": "Param",
            "type": "string"
          }
        },
        "required": [
          "node",
          "param"
        ],
        "title": "PromotedDoc",
        "type": "object"
      },
...
      "SubgraphDoc": {
        "additionalProperties": true,
        "description": "A reusable body. Instances are nodes typed ``subgraph:<id>``.\n\n``inputs``/``outputs`` name the boundary ports; ``promoted`` lists the inner params an\ninstance may override through its own ``params`` (keyed by ``\"<inner node>.<param>\"``) or\nlink to an input port through its own ``linked``.",
        "properties": {
          "edges": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EdgeDoc"
            },
            "title": "Edges",
            "type": "object"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/SubgraphPort"
            },
            "title": "Inputs",
            "type": "array"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "nodes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NodeDoc"
            },
            "title": "Nodes",
            "type": "object"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/SubgraphPort"
            },
            "title": "Outputs",
            "type": "array"
          },
          "promoted": {
            "items": {
              "$ref": "#/components/schemas/PromotedDoc"
            },
            "title": "Promoted",
            "type": "array"
          }
        },
        "title": "SubgraphDoc",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ViewDoc": {
        "additionalProperties": true,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "port": {
            "title": "Port",
            "type": "string"
          }
        },
        "required": [
          "id",
          "node",
          "port"
        ],
        "title": "ViewDoc",
        "type": "object"
      },
      "WorkflowDoc": {
        "additionalProperties": true,
        "description": "The canvas document. Unknown top-level fields are preserved.",
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "edges": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EdgeDoc"
            },
            "title": "Edges",
            "type": "object"
          },
          "format": {
            "const": "astro-canvas/workflow",
            "default": "astro-canvas/workflow",
            "title": "Format",
            "type": "string"
          },
          "groups": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GroupDoc"
            },
            "title": "Groups",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "layouts": {
            "additionalProperties": true,
            "title": "Layouts",
            "type": "object"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "name": {
            "default": "Untitled",
            "title": "Name",
            "type": "string"
          },
          "nodes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NodeDoc"
            },
            "title": "Nodes",
            "type": "object"
          },
          "promoted": {
            "items": {
              "$ref": "#/components/schemas/PromotedDoc"
            },
            "title": "Promoted",
            "type": "array"
          },
          "requires": {
            "additionalProperties": true,
            "title": "Requires",
            "type": "object"
          },
          "subgraphs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SubgraphDoc"
            },
            "title": "Subgraphs",
            "type": "object"
          },
          "version": {
            "const": 1,
            "default": 1,
            "title": "Version",
            "type": "integer"
          },
          "views": {
            "items": {
              "$ref": "#/components/schemas/ViewDoc"
            },
            "title": "Views",
            "type": "array"
          }
        },
        "title": "WorkflowDoc",
        "type": "object"
      },
      "WorkflowSaved": {
        "description": "A stored document plus the compile result for its current content.",
        "properties": {
          "doc": {
            "$ref": "#/components/schemas/WorkflowDoc"
          },
          "layout_errors": {
            "items": {
              "$ref": "#/components/schemas/LayoutIssue"
            },
            "title": "Layout Errors",
            "type": "array"
          },
          "node_errors": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/NodeIssue"
              },
              "type": "array"
            },
            "title": "Node Errors",
            "type": "object"
          }
        },
        "required": [
          "doc",
          "node_errors"
        ],
        "title": "WorkflowSaved",
        "type": "object"
      },
      "WorkflowSettings": {
        "description": "Per-workflow scheduler switches (kept in memory for the server's lifetime).",
        "properties": {
          "auto_run": {
            "default": true,
            "title": "Auto Run",
            "type": "boolean"
          }
        },
        "title": "WorkflowSettings",
        "type": "object"
      },
      "WorkflowStatus": {
        "properties": {
          "auto_run": {
            "default": true,
            "title": "Auto Run",
            "type": "boolean"
          },
          "current_run": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Run"
          },
          "node_errors": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/NodeIssue"
              },
              "type": "array"
            },
            "title": "Node Errors",
            "type": "object"
          },
          "nodes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NodeStatus"
            },
            "title": "Nodes",
            "type": "object"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "node_errors",
          "nodes"
        ],
        "title": "WorkflowStatus",
        "type": "object"
      },
      "WorkflowSummary": {
        "properties": {
          "created": {
            "title": "Created",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "modified": {
            "title": "Modified",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "node_count": {
            "title": "Node Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "created",
          "modified",
          "node_count",
          "hash"
        ],
        "title": "WorkflowSummary",
        "type": "object"
      },
      "WorkflowVersionInfo": {
        "properties": {
          "created": {
            "title": "Created",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "required": [
          "id",
          "created"
        ],
        "title": "WorkflowVersionInfo",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Astro Canvas",
    "version": "0.1.0a0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workflows": {
      "get": {
        "description": "Stored workflows, most recently modified first.",
        "operationId": "list_workflows_api_workflows_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowSummary"
                  },
                  "title": "Response List Workflows Api Workflows Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Workflows",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "description": "Store a new document (a fresh id is assigned when the given one already exists).",
        "operationId": "create_workflow_api_workflows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDoc"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{workflow_id}": {
      "delete": {
        "operationId": "delete_workflow_api_workflows__workflow_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Workflow",
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "get_workflow_api_workflows__workflow_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDoc"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow",
        "tags": [
          "workflows"
        ]
      },
      "put": {
        "description": "Replace the document; the engine recompiles it and auto-runs cheap dirty nodes.",
        "operationId": "put_workflow_api_workflows__workflow_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowDoc"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/api/workflows/{workflow_id}/settings": {
      "get": {
        "description": "The scheduler switches for one workflow.",
        "operationId": "get_settings_api_workflows__workflow_id__settings_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "description": "Toggle auto-run; enabling it schedules any dirty cheap nodes right away.",
        "operationId": "update_settings_api_workflows__workflow_id__settings_post",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowSettings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{workflow_id}/status": {
      "get": {
        "description": "Compile issues and the state of every node (what a reconnecting client needs).",
        "operationId": "workflow_status_api_workflows__workflow_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workflow Status",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/api/workflows/{workflow_id}/versions": {
      "get": {
        "description": "Auto-snapshots (every changed save) and labelled versions, newest first.",
        "operationId": "list_versions_api_workflows__workflow_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkflowVersionInfo"
                  },
                  "title": "Response List Versions Api Workflows  Workflow Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Versions",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflows/{workflow_id}/versions/{version_id}": {
      "get": {
        "operationId": "get_version_api_workflows__workflow_id__versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDoc"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Version",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}